refactor(connect): move core method definitions to corresponding folder
