+++ packages/connect-common/src/types/api/core/on.ts
@@ -1,0 +1,3 @@
+import type { ConnectEventCallbacks } from '../../emitter';
+
+export declare const on: ConnectEventCallbacks;
