+++ packages/connect-common/src/types/api/core/removeAllListeners.ts
@@ -1,3 +1,3 @@
-import type { ConnectEvents } from '../emitter';
+import type { ConnectEvents } from '../../emitter';
 
 export declare const removeAllListeners: (type?: ConnectEvents) => void;
