+++ packages/connect-common/src/types/api/core/uiResponse.ts
@@ -1,3 +1,3 @@
-import type { UiResponseEvent } from '../../events/ui-response';
+import type { UiResponseEvent } from '../../../events/ui-response';
 
 export declare function uiResponse(response: UiResponseEvent): void;
