--- packages/connect-common/src/types/api/on.ts
@@ -1,3 +1,0 @@
-import type { ConnectEventCallbacks } from '../emitter';
-
-export declare const on: ConnectEventCallbacks;
